refactor repeated questions into a concern
